% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_diagnosis_admission.R
\name{Extract_diagnosis_admission}
\alias{Extract_diagnosis_admission}
\title{Make a diagnosis table at admission}
\usage{
Extract_diagnosis_admission(data, diagnosis)
}
\arguments{
\item{data}{a spark data frame}

\item{diagnosis}{a character (list of character vector) specifying the condition(s) of interest}
}
\value{

}
\description{
Make a diagnosis table at admission
}
