% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_procedure_admission.R
\name{Extract_procedure_admission}
\alias{Extract_procedure_admission}
\title{Make a procedure table at admission}
\usage{
Extract_procedure_admission(data, procedure)
}
\arguments{
\item{data}{a spark data frame}

\item{diagnosis}{a character (list of character vector) specifying the type of procedure(s) of interest}
}
\value{

}
\description{
Make a procedure table at admission
}
