% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datalake.R
\name{read_datalake}
\alias{read_datalake}
\title{Read a tabular data file into a Spark DataFram from datalake storage account}
\usage{
read_datalake(filename, folder)
}
\arguments{
\item{filename}{The name of the data file (with its extension) to be imported}

\item{folder}{The name of the folder with in the storage account that holds 'filename'}
}
\value{
A Spark DataFrame object
}
\description{
Read a tabular data file into a Spark DataFram from datalake storage account
}
\examples{
\dontrun{
Inpatient_Demographic <- read_datalake("Inpatient_Demographic.csv", "Inpatient")}

}
