% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_categorical.R
\name{tab_categorical}
\alias{tab_categorical}
\title{Tabulate categorical data}
\usage{
tab_categorical(df, ...)
}
\arguments{
\item{df}{a spark data frame object}

\item{...}{}
}
\value{

}
\description{
Tabulate categorical data
}
